Improvements on grappling hook
removed issue of it shifting closer and closer to 90 on flipping, also
disabled changing length of the grappling hook whilst not connected.
This is to prevent the grappling hook from changing length incorrectly
when the player is moving. This is caused due to the velocity of the
player somehow having an impact. May wish to change this later if I can
figure out why.