did all of the things
loads of work on the level
added a non grapplable level
multiple new blocks
clouds stuff
background stuff
removed a load of rotation stuff from GrapplingEnd - this is to allow
the use of OnTriggerEnter2D instead of OnCollisionEnter2D (allows a
trigger being used), If rotation is later deemed to be necessary, will
have to use raycasting. Will be difficult.